cleanup: remove unused header files and reorganize
